-Separated Client class to Client and EvaluationInfo -Separated Client_UT to Client_UT and EvaluationInfo_UT -Created User, Manager, Physical Therapy class
